redoing in true tdd

+++ C/include/Malloc.h
@@ -1,6 +1,6 @@
 #ifndef M_CORE_C
 #define M_CORE_C
 #include <stdlib.h>
-typedef void* (*MallocFunc)(size_t size);
+typedef void* (*MallocFunc)(int size);
 extern MallocFunc Malloc;
 #endif
